Moved Scripting to its own package
